Вынес функции для сортировки в отдельный файл sort
