+++ src/utils/event.js
@@ -43,27 +43,4 @@ function isEventExpired (event) {
   return event.dateTo && dayjs().isAfter(event.dateTo, 'D');
 }
 
-function sortByDay(dayA, dayB) {
-  return dayjs(dayA.dateFrom).diff(dayjs(dayB.dateFrom));
-}
-
-function sortByEvent(eventA, eventB) {
-  return eventA.type.localeCompare(eventB.type);
-}
-
-function sortByTime(timeA, timeB){
-  const durationA = dayjs(timeA.dateFrom).diff(dayjs(timeA.dateTo));
-  const durationB = dayjs(timeB.dateFrom).diff(dayjs(timeB.dateTo));
-  return dayjs(durationA).diff(durationB);
-}
-
-function sortByPrice(priceA, priceB) {
-  return priceB.basePrice - priceA.basePrice;
-}
-
-function sortByOffers(offersA, offersB) {
-  return offersB.offers.length - offersA.offers.length;
-}
-
-
-export { humanizeDate, getTimeDifference, isEventFuture, isEventExpired, isEventActive, sortByDay, sortByEvent, sortByTime, sortByPrice, sortByOffers };
+export { humanizeDate, getTimeDifference, isEventFuture, isEventExpired, isEventActive};
